# packages/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Self Deepsearch Platform API
  version: 0.1.0
  description: Release A contract baseline. Automatic collection is not enabled.
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /admin/v1/imports/works/preflight:
    post:
      operationId: adminPreflightWorkCSV
      description: Parses and validates a work CSV without writing catalog data. Requires editor, admin or owner.
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "200":
          description: File-level and row-level validation report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkCSVPreflightResponse" }
        "400": { description: Missing file or malformed multipart body }
        "403": { description: Role is not allowed }
        "413": { description: File is larger than 1 MiB }
        "415": { description: Request is not multipart/form-data }
  /admin/v1/imports/works:
    get:
      operationId: adminListWorkImports
      description: Lists the latest 100 manual work CSV batches. Requires editor, admin or owner.
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Manual work import batches
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkImportBatchListResponse" }
    post:
      operationId: adminImportWorkCSV
      description: Atomically creates works, reviewing revisions, review tasks and provenance records. Repeating the same key and body returns the stored batch; a different body with the same key returns 409. Requires editor, admin or owner.
      security: [{ sessionCookie: [] }]
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, minLength: 8, maxLength: 128, pattern: "^[A-Za-z0-9._:-]+$" }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, reason]
              properties:
                file: { type: string, format: binary }
                reason: { type: string, minLength: 2, maxLength: 1000 }
      responses:
        "201":
          description: Accepted batch or idempotent replay
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkImportBatchResponse" }
        "409": { description: Database conflict or idempotency key reused with different content }
        "413": { description: File is larger than 1 MiB }
        "422":
          description: CSV has file-level or row-level validation errors
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkCSVPreflightResponse" }
...
components:
  securitySchemes:
    metricsBearer:
      type: http
      scheme: bearer
    sessionCookie:
      type: apiKey
      in: cookie
      name: sd_session
    recentAuthCookie:
      type: apiKey
      in: cookie
      name: sd_reauth
...
  schemas:
...
    CSVValidationIssue:
      type: object
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
    WorkCSVRow:
      type: object
      required: [row_number, code, title, issues]
      properties:
        row_number: { type: integer, minimum: 2 }
        code: { type: string }
        title: { type: string }
        title_original: { type: string }
        release_date: { type: string }
        studio_id: { type: string }
        performer_ids: { type: array, items: { type: string } }
        summary: { type: string }
        issues: { type: array, items: { $ref: "#/components/schemas/CSVValidationIssue" } }
    WorkCSVPreflightResponse:
      type: object
      required: [rows, file_issues, total_count, valid_count, invalid_count]
      properties:
        rows: { type: array, maxItems: 500, items: { $ref: "#/components/schemas/WorkCSVRow" } }
        file_issues: { type: array, items: { $ref: "#/components/schemas/CSVValidationIssue" } }
        total_count: { type: integer, minimum: 0, maximum: 500 }
        valid_count: { type: integer, minimum: 0, maximum: 500 }
        invalid_count: { type: integer, minimum: 0, maximum: 500 }
    WorkImportBatchResponse:
      type: object
      required: [id, status, input_count, accepted_count, rejected_count, entity_ids, idempotent_replay, received_at, completed_at]
      properties:
        id: { type: string, format: uuid }
        status: { const: accepted }
        input_count: { type: integer, minimum: 1, maximum: 500 }
        accepted_count: { type: integer, minimum: 1, maximum: 500 }
        rejected_count: { type: integer, minimum: 0, maximum: 500 }
        entity_ids: { type: array, items: { type: string, format: uuid } }
        idempotent_replay: { type: boolean }
        received_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
    WorkImportBatchListResponse:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/WorkImportBatchResponse" } }
...